Remove src folder
