+++ packages/transpiler/test/solc-0.5/test/compile.test.ts
@@ -1,4 +1,4 @@
-import { transpileAndSaveContracts, compileContracts } from '../src/index';
+import { transpileAndSaveContracts, compileContracts } from './setup';
 
 describe('When all the contracts transpilied and saved to contracts folder', (): void => {
   beforeAll(
